codeintel: Co-locate code next to background job
