+++ internal/codeintel/autoindexing/background/scheduler/iface.go
@@ -7,6 +7,15 @@ import (
 )
 
 type AutoIndexingService interface {
-	NewScheduler(interval time.Duration, repositoryProcessDelay time.Duration, repositoryBatchSize int, policyBatchSize int) goroutine.BackgroundRoutine
-	NewOnDemandScheduler(interval time.Duration, batchSize int) goroutine.BackgroundRoutine
+	NewScheduler(
+		interval time.Duration,
+		repositoryProcessDelay time.Duration,
+		repositoryBatchSize int,
+		policyBatchSize int,
+	) goroutine.BackgroundRoutine
+
+	NewOnDemandScheduler(
+		interval time.Duration,
+		batchSize int,
+	) goroutine.BackgroundRoutine
 }
