+++ internal/codeintel/autoindexing/background/scheduler/init.go
@@ -6,7 +6,16 @@ import (
 
 func NewSchedulers(autoIndexingSvc AutoIndexingService) []goroutine.BackgroundRoutine {
 	return []goroutine.BackgroundRoutine{
-		autoIndexingSvc.NewScheduler(ConfigInst.SchedulerInterval, ConfigInst.RepositoryProcessDelay, ConfigInst.RepositoryBatchSize, ConfigInst.PolicyBatchSize),
-		autoIndexingSvc.NewOnDemandScheduler(ConfigInst.OnDemandSchedulerInterval, ConfigInst.OnDemandBatchsize),
+		autoIndexingSvc.NewScheduler(
+			ConfigInst.SchedulerInterval,
+			ConfigInst.RepositoryProcessDelay,
+			ConfigInst.RepositoryBatchSize,
+			ConfigInst.PolicyBatchSize,
+		),
+
+		autoIndexingSvc.NewOnDemandScheduler(
+			ConfigInst.OnDemandSchedulerInterval,
+			ConfigInst.OnDemandBatchsize,
+		),
 	}
 }
