+++ internal/codeintel/dependencies/background/cratesyncer/config.go
@@ -1,0 +1,14 @@
+package cratesyncer
+
+import (
+	"github.com/sourcegraph/sourcegraph/internal/env"
+)
+
+type config struct {
+	env.BaseConfig
+}
+
+var ConfigInst = &config{}
+
+func (c *config) Load() {
+}
