+++ internal/codeintel/dependencies/background/cratesyncer/init.go
@@ -2,7 +2,7 @@ package cratesyncer
 
 import "github.com/sourcegraph/sourcegraph/internal/goroutine"
 
-func NewSyncers(depsSvc DependenciesService) []goroutine.BackgroundRoutine {
+func NewCrateSyncer(depsSvc DependenciesService) []goroutine.BackgroundRoutine {
 	return []goroutine.BackgroundRoutine{
 		depsSvc.NewCrateSyncer(),
 	}
