+++ internal/codeintel/uploads/background/expiration/init.go
@@ -6,7 +6,8 @@ import (
 
 func NewExpirationTasks(uploadSvc UploadService) []goroutine.BackgroundRoutine {
 	return []goroutine.BackgroundRoutine{
-		uploadSvc.NewExpirer(ConfigInst.ExpirerInterval,
+		uploadSvc.NewUploadExpirer(
+			ConfigInst.ExpirerInterval,
 			ConfigInst.RepositoryProcessDelay,
 			ConfigInst.RepositoryBatchSize,
 			ConfigInst.UploadProcessDelay,
@@ -14,6 +15,9 @@ func NewExpirationTasks(uploadSvc UploadService) []goroutine.BackgroundRoutine {
 			ConfigInst.CommitBatchSize,
 			ConfigInst.PolicyBatchSize,
 		),
-		uploadSvc.NewReferenceCountUpdater(ConfigInst.ReferenceCountUpdaterInterval, ConfigInst.ReferenceCountUpdaterBatchSize),
+		uploadSvc.NewReferenceCountUpdater(
+			ConfigInst.ReferenceCountUpdaterInterval,
+			ConfigInst.ReferenceCountUpdaterBatchSize,
+		),
 	}
 }
