--- internal/codeintel/uploads/refcount.go
@@ -1,14 +1,0 @@
-package uploads
-
-import (
-	"context"
-	"time"
-
-	"github.com/sourcegraph/sourcegraph/internal/goroutine"
-)
-
-func (s *Service) NewReferenceCountUpdater(interval time.Duration, batchSize int) goroutine.BackgroundRoutine {
-	return goroutine.NewPeriodicGoroutine(context.Background(), interval, goroutine.HandlerFunc(func(ctx context.Context) error {
-		return s.BackfillReferenceCountBatch(ctx, batchSize)
-	}))
-}
