+++ internal/codeintel/uploads/service.go
@@ -16,7 +16,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/codeintel/uploads/internal/store"
 	"github.com/sourcegraph/sourcegraph/internal/codeintel/uploads/shared"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
-	gitserverOptions "github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
 	"github.com/sourcegraph/sourcegraph/internal/workerutil/dbworker"
@@ -42,7 +41,6 @@ type service interface {
 	GetRepositoriesMaxStaleAge(ctx context.Context) (_ time.Duration, err error)
 	GetDirtyRepositories(ctx context.Context) (_ map[int]int, err error)
 	SetRepositoryAsDirty(ctx context.Context, repositoryID int) (err error)
-	UpdateDirtyRepositories(ctx context.Context, maxAgeForNonStaleBranches time.Duration, maxAgeForNonStaleTags time.Duration) (err error)
 	SetRepositoriesForRetentionScan(ctx context.Context, processDelay time.Duration, limit int) (_ []int, err error)
 
 	// Uploads
@@ -164,24 +162,6 @@ func (s *Service) GetCommitGraphMetadata(ctx context.Context, repositoryID int) 
 	return s.store.GetCommitGraphMetadata(ctx, repositoryID)
 }
 
-func (s *Service) UpdateSourcedCommits(ctx context.Context, repositoryID int, commit string, now time.Time) (uploadsUpdated int, err error) {
-	ctx, _, endObservation := s.operations.updateSourcedCommits.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{log.Int("repositoryID", repositoryID), log.String("commit", commit), log.String("now", now.String())},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.UpdateSourcedCommits(ctx, repositoryID, commit, now)
-}
-
-func (s *Service) DeleteSourcedCommits(ctx context.Context, repositoryID int, commit string, maximumCommitLag time.Duration, now time.Time) (uploadsUpdated int, uploadsDeleted int, err error) {
-	ctx, _, endObservation := s.operations.deleteSourcedCommits.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{log.Int("repositoryID", repositoryID), log.String("commit", commit), log.Int("maximumCommitLag in ms", int(maximumCommitLag.Milliseconds())), log.String("now", now.String())},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.DeleteSourcedCommits(ctx, repositoryID, commit, maximumCommitLag, now)
-}
-
 func (s *Service) GetOldestCommitDate(ctx context.Context, repositoryID int) (time.Time, bool, error) {
 	ctx, _, endObservation := s.operations.getOldestCommitDate.With(ctx, nil, observation.Args{
 		LogFields: []log.Field{log.Int("repositoryID", repositoryID)},
@@ -198,130 +178,6 @@ func (s *Service) SetRepositoryAsDirty(ctx context.Context, repositoryID int) (e
 	defer endObservation(1, observation.Args{})
 
 	return s.store.SetRepositoryAsDirty(ctx, repositoryID)
-}
-
-func (s *Service) UpdateDirtyRepositories(ctx context.Context, maxAgeForNonStaleBranches time.Duration, maxAgeForNonStaleTags time.Duration) (err error) {
-	ctx, _, endObservation := s.operations.updateDirtyRepositories.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{
-			log.Int("maxAgeForNonStaleBranches in ms", int(maxAgeForNonStaleBranches.Milliseconds())),
-			log.Int("maxAgeForNonStaleTags in ms", int(maxAgeForNonStaleTags.Milliseconds())),
-		},
-	})
-	defer endObservation(1, observation.Args{})
-
-	repositoryIDs, err := s.GetDirtyRepositories(ctx)
-	if err != nil {
-		return errors.Wrap(err, "uploadSvc.DirtyRepositories")
-	}
-
-	var updateErr error
-	for repositoryID, dirtyFlag := range repositoryIDs {
-		if err := s.lockAndUpdateUploadsVisibleToCommits(ctx, repositoryID, dirtyFlag, maxAgeForNonStaleBranches, maxAgeForNonStaleTags); err != nil {
-			if updateErr == nil {
-				updateErr = err
-			} else {
-				updateErr = errors.Append(updateErr, err)
-			}
-		}
-	}
-
-	return updateErr
-}
-
-// lockAndUpdateUploadsVisibleToCommits will call UpdateUploadsVisibleToCommits while holding an advisory lock to give exclusive access to the
-// update procedure for this repository. If the lock is already held, this method will simply do nothing.
-func (s *Service) lockAndUpdateUploadsVisibleToCommits(ctx context.Context, repositoryID, dirtyToken int, maxAgeForNonStaleBranches time.Duration, maxAgeForNonStaleTags time.Duration) (err error) {
-	ctx, trace, endObservation := s.operations.updateUploadsVisibleToCommits.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{
-			log.Int("repositoryID", repositoryID),
-			log.Int("dirtyToken", dirtyToken),
-		},
-	})
-	defer endObservation(1, observation.Args{})
-
-	ok, unlock, err := s.locker.Lock(ctx, int32(repositoryID), false)
-	if err != nil || !ok {
-		return errors.Wrap(err, "locker.Lock")
-	}
-	defer func() {
-		err = unlock(err)
-	}()
-
-	// The following process pulls the commit graph for the given repository from gitserver, pulls the set of LSIF
-	// upload objects for the given repository from Postgres, and correlates them into a visibility
-	// graph. This graph is then upserted back into Postgres for use by find closest dumps queries.
-	//
-	// The user should supply a dirty token that is associated with the given repository so that
-	// the repository can be unmarked as long as the repository is not marked as dirty again before
-	// the update completes.
-
-	// Construct a view of the git graph that we will later decorate with upload information.
-	commitGraph, err := s.getCommitGraph(ctx, repositoryID)
-	if err != nil {
-		return err
-	}
-	trace.Log(log.Int("numCommitGraphKeys", len(commitGraph.Order())))
-
-	refDescriptions, err := s.gitserverClient.RefDescriptions(ctx, repositoryID)
-	if err != nil {
-		return errors.Wrap(err, "gitserver.RefDescriptions")
-	}
-	trace.Log(log.Int("numRefDescriptions", len(refDescriptions)))
-
-	// Decorate the commit graph with the set of processed uploads are visible from each commit,
-	// then bulk update the denormalized view in Postgres. We call this with an empty graph as well
-	// so that we end up clearing the stale data and bulk inserting nothing.
-	if err := s.UpdateUploadsVisibleToCommits(ctx, repositoryID, commitGraph, refDescriptions, maxAgeForNonStaleBranches, maxAgeForNonStaleTags, dirtyToken, time.Time{}); err != nil {
-		return errors.Wrap(err, "uploadSvc.UpdateUploadsVisibleToCommits")
-	}
-
-	return nil
-}
-
-// getCommitGraph builds a partial commit graph that includes the most recent commits on each branch
-// extending back as as the date of the oldest commit for which we have a processed upload for this
-// repository.
-//
-// This optimization is necessary as decorating the commit graph is an operation that scales with
-// the size of both the git graph and the number of uploads (multiplicatively). For repositories with
-// a very large number of commits or distinct roots (most monorepos) this is a necessary optimization.
-//
-// The number of commits pulled back here should not grow over time unless the repo is growing at an
-// accelerating rate, as we routinely expire old information for active repositories in a janitor
-// process.
-func (s *Service) getCommitGraph(ctx context.Context, repositoryID int) (*gitdomain.CommitGraph, error) {
-	commitDate, ok, err := s.GetOldestCommitDate(ctx, repositoryID)
-	if err != nil {
-		return nil, err
-	}
-	if !ok {
-		// No uploads exist for this repository
-		return gitdomain.ParseCommitGraph(nil), nil
-	}
-
-	// The --since flag for git log is exclusive, but we want to include the commit where the
-	// oldest dump is defined. This flag only has second resolution, so we shouldn't be pulling
-	// back any more data than we wanted.
-	commitDate = commitDate.Add(-time.Second)
-
-	commitGraph, err := s.gitserverClient.CommitGraph(ctx, repositoryID, gitserverOptions.CommitGraphOptions{
-		AllRefs: true,
-		Since:   &commitDate,
-	})
-	if err != nil {
-		return nil, errors.Wrap(err, "gitserver.CommitGraph")
-	}
-
-	return commitGraph, nil
-}
-
-func (s *Service) SetRepositoriesForRetentionScan(ctx context.Context, processDelay time.Duration, limit int) (_ []int, err error) {
-	ctx, _, endObservation := s.operations.setRepositoriesForRetentionScan.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{log.Int("processDelay in ms", int(processDelay.Milliseconds())), log.Int("limit", limit)},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.SetRepositoriesForRetentionScan(ctx, processDelay, limit)
 }
 
 func (s *Service) GetRepoName(ctx context.Context, repositoryID int) (_ string, err error) {
@@ -352,13 +208,6 @@ func (s *Service) GetRepositoriesMaxStaleAge(ctx context.Context) (_ time.Durati
 	defer endObservation(1, observation.Args{})
 
 	return s.store.GetRepositoriesMaxStaleAge(ctx)
-}
-
-func (s *Service) GetDirtyRepositories(ctx context.Context) (_ map[int]int, err error) {
-	ctx, _, endObservation := s.operations.getDirtyRepositories.With(ctx, &err, observation.Args{})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.GetDirtyRepositories(ctx)
 }
 
 func (s *Service) GetUploads(ctx context.Context, opts types.GetUploadsOptions) (uploads []types.Upload, totalCount int, err error) {
@@ -401,23 +250,6 @@ func (s *Service) GetUploadIDsWithReferences(ctx context.Context, orderedMoniker
 	return s.store.GetUploadIDsWithReferences(ctx, orderedMonikers, ignoreIDs, repositoryID, commit, limit, offset, trace)
 }
 
-func (s *Service) UpdateUploadsVisibleToCommits(ctx context.Context, repositoryID int, graph *gitdomain.CommitGraph, refDescriptions map[string][]gitdomain.RefDescription, maxAgeForNonStaleBranches, maxAgeForNonStaleTags time.Duration, dirtyToken int, now time.Time) (err error) {
-	ctx, _, endObservation := s.operations.updateUploadsVisibleToCommits.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{
-			log.Int("repositoryID", repositoryID),
-			log.String("graph", fmt.Sprintf("%v", graph)),
-			log.String("refDescriptions", fmt.Sprintf("%v", refDescriptions)),
-			log.String("maxAgeForNonStaleBranches", maxAgeForNonStaleBranches.String()),
-			log.String("maxAgeForNonStaleTags", maxAgeForNonStaleTags.String()),
-			log.Int("dirtyToken", dirtyToken),
-			log.String("now", now.String()),
-		},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.UpdateUploadsVisibleToCommits(ctx, repositoryID, graph, refDescriptions, maxAgeForNonStaleBranches, maxAgeForNonStaleTags, dirtyToken, now)
-}
-
 func (s *Service) UpdateUploadRetention(ctx context.Context, protectedIDs, expiredIDs []int) (err error) {
 	ctx, _, endObservation := s.operations.updateUploadRetention.With(ctx, &err, observation.Args{
 		LogFields: []log.Field{log.String("protectedIDs", fmt.Sprintf("%v", protectedIDs)), log.String("expiredIDs", fmt.Sprintf("%v", expiredIDs))},
@@ -426,16 +258,6 @@ func (s *Service) UpdateUploadRetention(ctx context.Context, protectedIDs, expir
 
 	return s.store.UpdateUploadRetention(ctx, protectedIDs, expiredIDs)
 }
-
-func (s *Service) BackfillReferenceCountBatch(ctx context.Context, batchSize int) (err error) {
-	ctx, _, endObservation := s.operations.backfillReferenceCountBatch.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{log.Int("batchSize", batchSize)},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.BackfillReferenceCountBatch(ctx, batchSize)
-}
-
 func (s *Service) UpdateUploadsReferenceCounts(ctx context.Context, ids []int, dependencyUpdateType shared.DependencyReferenceCountUpdateType) (updated int, err error) {
 	ctx, _, endObservation := s.operations.updateUploadsReferenceCounts.With(ctx, &err, observation.Args{
 		LogFields: []log.Field{log.String("ids", fmt.Sprintf("%v", ids))},
@@ -443,22 +265,6 @@ func (s *Service) UpdateUploadsReferenceCounts(ctx context.Context, ids []int, d
 	defer endObservation(1, observation.Args{})
 
 	return s.store.UpdateUploadsReferenceCounts(ctx, ids, dependencyUpdateType)
-}
-
-func (s *Service) SoftDeleteExpiredUploads(ctx context.Context) (count int, err error) {
-	ctx, _, endObservation := s.operations.softDeleteExpiredUploads.With(ctx, &err, observation.Args{})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.SoftDeleteExpiredUploads(ctx)
-}
-
-func (s *Service) DeleteUploadsStuckUploading(ctx context.Context, uploadedBefore time.Time) (_ int, err error) {
-	ctx, _, endObservation := s.operations.deleteUploadsStuckUploading.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{log.String("uploadedBefore", uploadedBefore.String())},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.DeleteUploadsStuckUploading(ctx, uploadedBefore)
 }
 
 func (s *Service) DeleteUploadsWithoutRepository(ctx context.Context, now time.Time) (_ map[int]int, err error) {
@@ -600,46 +406,6 @@ func (s *Service) GetDumpsByIDs(ctx context.Context, ids []int) (_ []types.Dump,
 	return s.store.GetDumpsByIDs(ctx, ids)
 }
 
-func (s *Service) HardDeleteExpiredUploads(ctx context.Context) (count int, err error) {
-	ctx, _, endObservation := s.operations.hardDeleteUploads.With(ctx, &err, observation.Args{})
-	defer endObservation(1, observation.Args{})
-
-	const uploadsBatchSize = 100
-	options := types.GetUploadsOptions{
-		State:            "deleted",
-		Limit:            uploadsBatchSize,
-		AllowExpired:     true,
-		AllowDeletedRepo: true,
-	}
-
-	for {
-		// Always request the first page of deleted uploads. If this is not
-		// the first iteration of the loop, then the previous iteration has
-		// deleted the records that composed the previous page, and the
-		// previous "second" page is now the first page.
-		uploads, totalCount, err := s.store.GetUploads(ctx, options)
-		if err != nil {
-			return 0, errors.Wrap(err, "store.GetUploads")
-		}
-
-		ids := uploadIDs(uploads)
-		if err := s.lsifstore.DeleteLsifDataByUploadIds(ctx, ids...); err != nil {
-			return 0, errors.Wrap(err, "lsifstore.Clear")
-		}
-
-		if err := s.store.HardDeleteUploadsByIDs(ctx, ids...); err != nil {
-			return 0, errors.Wrap(err, "store.HardDeleteUploadsByIDs")
-		}
-
-		count += len(uploads)
-		if count >= totalCount {
-			break
-		}
-	}
-
-	return count, nil
-}
-
 func (s *Service) UpdatePackages(ctx context.Context, dumpID int, packages []precise.Package) (err error) {
 	ctx, _, endObservation := s.operations.updatePackages.With(ctx, &err, observation.Args{
 		LogFields: []log.Field{log.Int("dumpID", dumpID), log.String("packages", fmt.Sprintf("%v", packages))},
@@ -676,56 +442,6 @@ func (s *Service) GetAuditLogsForUpload(ctx context.Context, uploadID int) (_ []
 	return s.store.GetAuditLogsForUpload(ctx, uploadID)
 }
 
-func (s *Service) DeleteOldAuditLogs(ctx context.Context, maxAge time.Duration, now time.Time) (count int, err error) {
-	ctx, _, endObservation := s.operations.deleteOldAuditLogs.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{log.String("maxAge", maxAge.String()), log.String("now", now.String())},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.DeleteOldAuditLogs(ctx, maxAge, now)
-}
-
-// BackfillCommittedAtBatch calculates the committed_at value for a batch of upload records that do not have
-// this value set. This method is used to backfill old upload records prior to this value being reliably set
-// during processing.
-func (s *Service) BackfillCommittedAtBatch(ctx context.Context, batchSize int) (err error) {
-	ctx, _, endObservation := s.operations.backfillCommittedAtBatch.With(ctx, &err, observation.Args{LogFields: []log.Field{
-		log.Int("batchSize", batchSize),
-	}})
-	defer endObservation(1, observation.Args{})
-
-	tx, err := s.store.Transact(ctx)
-	defer func() {
-		err = tx.Done(err)
-	}()
-
-	batch, err := tx.SourcedCommitsWithoutCommittedAt(ctx, batchSize)
-	if err != nil {
-		return errors.Wrap(err, "store.SourcedCommitsWithoutCommittedAt")
-	}
-
-	for _, sourcedCommits := range batch {
-		for _, commit := range sourcedCommits.Commits {
-			commitDateString, err := s.getCommitDate(ctx, sourcedCommits.RepositoryID, commit)
-			if err != nil {
-				return err
-			}
-
-			// Update commit date of all uploads attached to this this repository and commit
-			if err := tx.UpdateCommittedAt(ctx, sourcedCommits.RepositoryID, commit, commitDateString); err != nil {
-				return errors.Wrap(err, "store.UpdateCommittedAt")
-			}
-		}
-
-		// Mark repository as dirty so the commit graph is recalculated with fresh data
-		if err := tx.SetRepositoryAsDirty(ctx, sourcedCommits.RepositoryID); err != nil {
-			return errors.Wrap(err, "store.SetRepositoryAsDirty")
-		}
-	}
-
-	return nil
-}
-
 func (s *Service) GetUploadDocumentsForPath(ctx context.Context, bundleID int, pathPattern string) (_ []string, _ int, err error) {
 	ctx, _, endObservation := s.operations.getUploadDocumentsForPath.With(ctx, &err, observation.Args{
 		LogFields: []log.Field{log.Int("bundleID", bundleID), log.String("pathPattern", pathPattern)},
@@ -751,26 +467,6 @@ func (s *Service) GetLastUploadRetentionScanForRepository(ctx context.Context, r
 	defer endObservation(1, observation.Args{})
 
 	return s.store.GetLastUploadRetentionScanForRepository(ctx, repositoryID)
-}
-
-func (s *Service) getCommitDate(ctx context.Context, repositoryID int, commit string) (string, error) {
-	_, commitDate, revisionExists, err := s.gitserverClient.CommitDate(ctx, repositoryID, commit)
-	if err != nil {
-		return "", errors.Wrap(err, "gitserver.CommitDate")
-	}
-
-	var commitDateString string
-	if revisionExists {
-		commitDateString = commitDate.Format(time.RFC3339)
-	} else {
-		// Set a value here that we'll filter out on the query side so that we don't
-		// reprocess the same failing batch infinitely. We could alternatively soft
-		// delete the record, but it would be better to keep record deletion behavior
-		// together in the same place (so we have unified metrics on that event).
-		commitDateString = "-infinity"
-	}
-
-	return commitDateString, nil
 }
 
 func (s *Service) GetListTags(ctx context.Context, repo api.RepoName, commitObjs ...string) (_ []*gitdomain.Tag, err error) {
